Use strict mode, add global variables to IIFE
Strict mode makes it easier to avoid syntax errors by throwing them
instead of failing silently. Since it's early in the project and it's not
complicated yet, starting using strict mode came with almost no cost.

Added arguments representing `window` and `document` object to IIFE
statement. Inside the IIFE they are referenced as `win` and `doc`
accordingly. This is just for convinience purposes.